livekit-agent/requirements.txt
# LiveKit Agents Framework
livekit>=0.17.0
livekit-agents>=0.10.0
livekit-plugins-tavus>=0.1.0

# Google Gemini API
google-generativeai>=0.8.0

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
aiohttp>=3.9.0
